+++ game/CMakeLists.txt
@@ -35,7 +35,6 @@ set(RUNTIME_SOURCE
         graphics/jak2_texture_remap.cpp
         graphics/opengl_renderer/background/background_common.cpp
         graphics/opengl_renderer/background/Shrub.cpp
-        graphics/opengl_renderer/background/Tfrag3.cpp
         graphics/opengl_renderer/background/TFragment.cpp
         graphics/opengl_renderer/background/Tie3.cpp
         graphics/opengl_renderer/BlitDisplays.cpp
