+++ game/graphics/opengl_renderer/OpenGLRenderer.cpp
@@ -102,8 +102,7 @@ OpenGLRenderer::OpenGLRenderer(std::shared_ptr<TexturePool> texture_pool,
       ASSERT(false);
   }
 
-  m_merc2 = std::make_shared<Merc2>(m_render_state.shaders,
-                                    m_texture_animator ? m_texture_animator->slots() : nullptr);
+  m_merc2 = std::make_shared<Merc2>(m_render_state.shaders, anim_slot_array());
   m_generic2 = std::make_shared<Generic2>(m_render_state.shaders);
 
   // initialize all renderers
@@ -143,7 +142,7 @@ void OpenGLRenderer::init_bucket_renderers_jak2() {
     init_bucket_renderer<TFragment>(
         fmt::format("tfrag-l{}-tfrag", i), BucketCategory::TFRAG,
         GET_BUCKET_ID_FOR_LIST(BucketId::TFRAG_L0_TFRAG, BucketId::TFRAG_L1_TFRAG, i),
-        std::vector{tfrag3::TFragmentTreeKind::NORMAL}, false, i);
+        std::vector{tfrag3::TFragmentTreeKind::NORMAL}, false, i, anim_slot_array());
     Tie3* tie = init_bucket_renderer<Tie3>(
         fmt::format("tie-l{}-tfrag", i), BucketCategory::TIE,
         GET_BUCKET_ID_FOR_LIST(BucketId::TIE_L0_TFRAG, BucketId::TIE_L1_TFRAG, i), i);
@@ -181,7 +180,7 @@ void OpenGLRenderer::init_bucket_renderers_jak2() {
     init_bucket_renderer<TFragment>(
         fmt::format("tfrag-t-l{}-alpha", i), BucketCategory::TFRAG,
         GET_BUCKET_ID_FOR_LIST(BucketId::TFRAG_T_L0_ALPHA, BucketId::TFRAG_T_L1_ALPHA, i),
-        std::vector{tfrag3::TFragmentTreeKind::TRANS}, false, i);
+        std::vector{tfrag3::TFragmentTreeKind::TRANS}, false, i, anim_slot_array());
     init_bucket_renderer<Tie3AnotherCategory>(
         fmt::format("tie-t-l{}-alpha", i), BucketCategory::TIE,
         GET_BUCKET_ID_FOR_LIST(BucketId::TIE_T_L0_ALPHA, BucketId::TIE_T_L1_ALPHA, i), tie,
@@ -236,7 +235,7 @@ void OpenGLRenderer::init_bucket_renderers_jak2() {
     init_bucket_renderer<TFragment>(
         fmt::format("tfrag-w-l{}-alpha", i), BucketCategory::TFRAG,
         GET_BUCKET_ID_FOR_LIST(BucketId::TFRAG_W_L0_WATER, BucketId::TFRAG_W_L1_WATER, i),
-        std::vector{tfrag3::TFragmentTreeKind::WATER}, false, i);
+        std::vector{tfrag3::TFragmentTreeKind::WATER}, false, i, anim_slot_array());
     init_bucket_renderer<Tie3AnotherCategory>(
         fmt::format("tie-w-l{}-water", i), BucketCategory::TIE,
         GET_BUCKET_ID_FOR_LIST(BucketId::TIE_W_L0_WATER, BucketId::TIE_W_L1_WATER, i), tie,
@@ -355,7 +354,7 @@ void OpenGLRenderer::init_bucket_renderers_jak1() {
                                              BucketId::TFRAG_TEX_LEVEL0, m_texture_animator);
   // 6 : TFRAG_LEVEL0
   init_bucket_renderer<TFragment>("l0-tfrag-tfrag", BucketCategory::TFRAG, BucketId::TFRAG_LEVEL0,
-                                  normal_tfrags, false, 0);
+                                  normal_tfrags, false, 0, anim_slot_array());
   // 7 : TFRAG_NEAR_LEVEL0
   // 8 : TIE_NEAR_LEVEL0
   // 9 : TIE_LEVEL0
@@ -377,7 +376,7 @@ void OpenGLRenderer::init_bucket_renderers_jak1() {
                                              BucketId::TFRAG_TEX_LEVEL1, m_texture_animator);
   // 13 : TFRAG_LEVEL1
   init_bucket_renderer<TFragment>("l1-tfrag-tfrag", BucketCategory::TFRAG, BucketId::TFRAG_LEVEL1,
-                                  normal_tfrags, false, 1);
+                                  normal_tfrags, false, 1, anim_slot_array());
   // 14 : TFRAG_NEAR_LEVEL1
   // 15 : TIE_NEAR_LEVEL1
   // 16 : TIE_LEVEL1
@@ -430,14 +429,15 @@ void OpenGLRenderer::init_bucket_renderers_jak1() {
                                              BucketId::ALPHA_TEX_LEVEL0, m_texture_animator);  // 31
   init_bucket_renderer<SkyBlendHandler>("l0-alpha-sky-blend-and-tfrag-trans", BucketCategory::OTHER,
                                         BucketId::TFRAG_TRANS0_AND_SKY_BLEND_LEVEL0, 0,
-                                        sky_gpu_blender, sky_cpu_blender);  // 32
+                                        sky_gpu_blender, sky_cpu_blender, anim_slot_array());  // 32
   // 33
   init_bucket_renderer<TFragment>("l0-alpha-tfrag", BucketCategory::TFRAG,
-                                  BucketId::TFRAG_DIRT_LEVEL0, dirt_tfrags, false,
-                                  0);  // 34
+                                  BucketId::TFRAG_DIRT_LEVEL0, dirt_tfrags, false, 0,
+                                  anim_slot_array());  // 34
   // 35
   init_bucket_renderer<TFragment>("l0-alpha-tfrag-ice", BucketCategory::TFRAG,
-                                  BucketId::TFRAG_ICE_LEVEL0, ice_tfrags, false, 0);
+                                  BucketId::TFRAG_ICE_LEVEL0, ice_tfrags, false, 0,
+                                  anim_slot_array());
   // 37
 
   //-----------------------
@@ -447,14 +447,15 @@ void OpenGLRenderer::init_bucket_renderers_jak1() {
                                              BucketId::ALPHA_TEX_LEVEL1, m_texture_animator);  // 38
   init_bucket_renderer<SkyBlendHandler>("l1-alpha-sky-blend-and-tfrag-trans", BucketCategory::OTHER,
                                         BucketId::TFRAG_TRANS1_AND_SKY_BLEND_LEVEL1, 1,
-                                        sky_gpu_blender, sky_cpu_blender);  // 39
+                                        sky_gpu_blender, sky_cpu_blender, anim_slot_array());  // 39
   // 40
   init_bucket_renderer<TFragment>("l1-alpha-tfrag-dirt", BucketCategory::TFRAG,
-                                  BucketId::TFRAG_DIRT_LEVEL1, dirt_tfrags, false,
-                                  1);  // 41
+                                  BucketId::TFRAG_DIRT_LEVEL1, dirt_tfrags, false, 1,
+                                  anim_slot_array());  // 41
   // 42
   init_bucket_renderer<TFragment>("l1-alpha-tfrag-ice", BucketCategory::TFRAG,
-                                  BucketId::TFRAG_ICE_LEVEL1, ice_tfrags, false, 1);
+                                  BucketId::TFRAG_ICE_LEVEL1, ice_tfrags, false, 1,
+                                  anim_slot_array());
   // 44
 
   init_bucket_renderer<Merc2BucketRenderer>("common-alpha-merc", BucketCategory::MERC,
