+++ goal_src/jak2/engine/gfx/texture/texture-anim.gc
@@ -659,12 +659,15 @@
      )
     ((= anim-array *skull-gem-texture-anim-array*)
      (pc-update-fixed-anim bucket (texture-anim-pc skull-gem) anim-array)
+     (return #f)
      )
     ((= anim-array *bomb-texture-anim-array*)
      (pc-update-fixed-anim bucket (texture-anim-pc bomb) anim-array)
+     (return #f)
      )
     ((= anim-array *cas-conveyor-texture-anim-array*)
      (pc-update-fixed-anim bucket (texture-anim-pc cas-conveyor) anim-array)
+     (return #f)
      )
     ((= anim-array *darkjak-texture-anim-array*)
      ;; darkjak is simple, and we reimplemented it in C++.
